Add files via upload
Key file is in 'Encoding ML project openrice' where I import the sentence transformers model (all-MiniLM-L6-v2) and encode the data from the restaurant guide (OpenRice). 
Cosine similarity search is performed on the test sentence and compared to the data in the OpenRice database, and prints out the top 10 results in order of similarity.